+++ src/interfaces/node-transformers/rename-properties-transformers/replacer/IRenamePropertiesReplacer.ts
@@ -7,8 +7,8 @@ export interface IRenamePropertiesReplacer {
     excludePropertyName(propertyName: string): void;
 
     /**
-     * @param {ESTree.Identifier | ESTree.Literal} node
-     * @returns {ESTree.Identifier | ESTree.Literal}
+     * @param {ESTree.Identifier | ESTree.Literal | ESTree.PrivateIdentifier} node
+     * @returns {ESTree.Identifier | ESTree.Literal | ESTree.PrivateIdentifier}
      */
-    replace(node: ESTree.Identifier | ESTree.Literal): ESTree.Identifier | ESTree.Literal;
+    replace(node: ESTree.Identifier | ESTree.Literal | ESTree.PrivateIdentifier): ESTree.Identifier | ESTree.Literal | ESTree.PrivateIdentifier;
 }